Create CARRHIM.c
In this problem , in every play we have to count simply white and black pawns covered. And then decide which team will be given white covers and which will be given black .The main thing is for winning a team that any team must have already covered red pawn but not necessarily winning team.