Create DIVISION.c
The main idea behind this problem is using modulus when a/b is not an integer then finding a/b%mod.where a=number whose binary form is of length n i.e (2^n-1) and for divisibility of 3 b=3.So we only have to find (2^n-1)/3%mod + 1(for 0 i.e. obviously divisible by 3).